Code refactor: move parameter parsing code into parameter.h/c
