+++ src/ntttcp.h
@@ -4,6 +4,9 @@
 // Author: Shihua (Simon) Xiao, sixiao@microsoft.com
 // ----------------------------------------------------------------------------------
 
+#define __STDC_FORMAT_MACROS
+#pragma once
+
 #define _GNU_SOURCE
 #include <stdbool.h>
 #include <stdlib.h>
